feat: LottoNumber 클래스 추가
