+++ src/main/java/domain/WinningNumbers.java
@@ -1,4 +1,4 @@
-package lotto;
+package domain;
 
 import java.util.ArrayList;
 import java.util.List;
