+++ src/test/java/domain/RankTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package domain;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
